Add third class progression

+++ database/sql/database.sql
@@ -20,7 +20,7 @@ CREATE TABLE `characters`(
     `maxHp`       float       NOT NULL,
     `mp`          float       NOT NULL DEFAULT 25,
     `maxMp`       float       NOT NULL,
-    `exp`         int(20)     NOT NULL DEFAULT 0,
+    `exp`         bigint      NOT NULL DEFAULT 0,
     `sp`          int(10)     NOT NULL DEFAULT 0,
     `pk`          int(10)     NOT NULL DEFAULT 0,
     `pvp`         int(10)     NOT NULL DEFAULT 0,
